2457: First pass at review points
