+++ native/src/components/OpeningEntry.tsx
@@ -1,91 +1,19 @@
 import React, { ReactElement, useState } from 'react'
-import { Trans, useTranslation } from 'react-i18next'
-import { Modal as RNModal, Text } from 'react-native'
+import { useTranslation } from 'react-i18next'
 import styled from 'styled-components/native'
 
 import { TimeSlot } from 'shared/api/types'
 
 import { NoteIcon } from '../assets'
-import { contentDirection } from '../constants/contentDirection'
-import Link from './Link'
+import { contentDirection, isContentDirectionReversalRequired } from '../constants/contentDirection'
+import AppointmentOnlyOverlay from './AppointmentOnlyOverlay'
 import Icon from './base/Icon'
-
-const BackgroundForClosing = styled.Pressable`
-  flex: 1;
-  background-color: ${props => props.theme.colors.textDecorationColor};
-  opacity: 0.25;
-`
-
-const OverlayCenterer = styled.View`
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  justify-content: center;
-  align-items: center;
-`
-
-const OverlayContainer = styled.View`
-  background-color: white;
-  margin: 32px;
-  padding: 24px;
-  border-radius: 28px;
-`
-
-const OverlayTitle = styled.Text`
-  font-weight: 700;
-  margin-bottom: 16px;
-`
-
-const CloseButton = styled.Pressable`
-  padding: 12px;
-  margin-left: auto;
-`
-
-const CloseButtonText = styled.Text`
-  font-weight: 700;
-  color: ${props => props.theme.colors.textSecondaryColor};
-`
-
-type AppointmentOnlyOverlayProps = {
-  link: string | null
-  closeOverlay: () => void
-}
-
-const AppointmentOnlyOverlay = ({ link, closeOverlay }: AppointmentOnlyOverlayProps): ReactElement => {
-  const { t } = useTranslation('pois')
-  return (
-    <RNModal onRequestClose={closeOverlay} animationType='fade' transparent>
-      <BackgroundForClosing onPress={closeOverlay} />
-      <OverlayCenterer>
-        <OverlayContainer>
-          <OverlayTitle>{t('appointmentNecessary')}</OverlayTitle>
-          <Text>
-            {link ? (
-              // More information: https://react.i18next.com/latest/trans-component
-              <Trans i18nKey='makeAppointmentTooltipWithLink' ns='pois'>
-                This gets replaced by react-18next
-                <Link url={link} text={t('theWebsite')} />
-              </Trans>
-            ) : (
-              t('makeAppointmentTooltipWithoutLink')
-            )}
-          </Text>
-          <CloseButton onPress={closeOverlay}>
-            <CloseButtonText>{t('close', { ns: 'common' })}</CloseButtonText>
-          </CloseButton>
-        </OverlayContainer>
-      </OverlayCenterer>
-    </RNModal>
-  )
-}
 
 const EntryContainer = styled.View<{ language: string }>`
   display: flex;
   flex-direction: ${props => contentDirection(props.language)};
   justify-content: space-between;
-  padding: ${props => (contentDirection(props.language) === 'row' ? '4px 36px 4px 0' : '4px 26px 4px 0')};
+  padding: ${props => (isContentDirectionReversalRequired(props.language) ? '4px 26px 4px 0' : '4px 36px 4px 0')};
 `
 
 const Timeslot = styled.View`
@@ -104,10 +32,10 @@ const TimeSlotLabel = styled.Text<{ isCurrentDay: boolean }>`
     props.isCurrentDay ? props.theme.fonts.native.contentFontBold : props.theme.fonts.native.contentFontRegular};
 `
 
-const AppointmentOnlyContainer = styled.View`
+const AppointmentOnlyContainer = styled.View<{ language: string }>`
   position: absolute;
-  right: -3px;
   top: 6px;
+  ${props => (isContentDirectionReversalRequired(props.language) ? 'right: 3px;' : 'right: -3px;')}
 `
 
 const StyledPressable = styled.Pressable`
@@ -128,7 +56,7 @@ type OpeningEntryProps = {
   isCurrentDay: boolean
   language: string
   appointmentOnly: boolean
-  link: string | null
+  appointmentOverlayLink: string | null
 }
 
 const OpeningEntry = ({
@@ -139,7 +67,7 @@ const OpeningEntry = ({
   isCurrentDay,
   language,
   appointmentOnly,
-  link,
+  appointmentOverlayLink,
 }: OpeningEntryProps): ReactElement => {
   const { t } = useTranslation('pois')
 
@@ -160,11 +88,16 @@ const OpeningEntry = ({
         </Timeslot>
       )}
       {appointmentOnly && (
-        <AppointmentOnlyContainer>
+        <AppointmentOnlyContainer language={language}>
           <StyledPressable onPress={() => setOverlayOpen(true)}>
             <StyledIcon Icon={NoteIcon} label={t('appointmentNecessary')} />
           </StyledPressable>
-          {overlayOpen && <AppointmentOnlyOverlay closeOverlay={() => setOverlayOpen(false)} link={link} />}
+          {overlayOpen && (
+            <AppointmentOnlyOverlay
+              closeOverlay={() => setOverlayOpen(false)}
+              appointmentUrl={appointmentOverlayLink}
+            />
+          )}
         </AppointmentOnlyContainer>
       )}
     </EntryContainer>
