+++ native/src/components/PoiDetails.tsx
@@ -105,7 +105,7 @@ const PoiDetails = ({ poi, language, distance }: PoiDetailsProps): ReactElement 
         isCurrentlyOpen={isCurrentlyOpen}
         isTemporarilyClosed={temporarilyClosed}
         appointmentUrl={appointmentUrl}
-        link={appointmentUrl || website}
+        appointmentOverlayLink={appointmentUrl || website}
       />
       {content.length > 0 && (
         <>
