+++ translations/translations.json
@@ -8145,8 +8145,7 @@
       "makeAppointment": "Termin vereinbaren",
       "appointmentNecessary": "Termin erforderlich",
       "makeAppointmentTooltipWithLink": "Vereinbare an diesem Tag einen Termin über <1>die Webseite</1> oder telefonisch.",
-      "theWebsite": "die Webseite",
-      "makeAppointmentTooltipWithoutLink": "Vereinbare an diesem Tag einen Termin über die Webseite oder telefonisch."
+      "theWebsite": "die Webseite"
     },
     "am": {
       "pageTitle": "ካርታ",
@@ -8556,8 +8555,7 @@
       "makeAppointment": "Make an appointment",
       "appointmentNecessary": "Appointment necessary",
       "makeAppointmentTooltipWithLink": "Make an appointment on this day via <1>the website</1> or by phone.",
-      "theWebsite": "the website",
-      "makeAppointmentTooltipWithoutLink": "Make an appointment on this day via the website or by phone."
+      "theWebsite": "the website"
     },
     "es": {
       "pageTitle": "Mapa",
