+++ web/src/components/PoiDetails.tsx
@@ -207,7 +207,7 @@ const PoiDetails = ({ poi, distance, toolbar }: PoiDetailsProps): ReactElement =
           openingHours={openingHours}
           isCurrentlyOpen={isCurrentlyOpen}
           isTemporarilyClosed={temporarilyClosed}
-          link={appointmentUrl ?? website}
+          appointmentOverlayLink={appointmentUrl ?? website}
         />
         {appointmentUrl !== null && (
           <Link to={appointmentUrl} newTab>
